toFloat price in all project

+++ app/src/main/java/team/four/mys/data/repository/FirebaseRepositoryImpl.kt
@@ -16,7 +16,7 @@ class FirebaseRepositoryImpl(private val firebaseDatabase: FirebaseDatabase) : F
         firebaseDatabase.deleteSubscription(deleteSubscriptionParam = deleteSubscriptionParam)
     }
 
-    override suspend fun getPriceFirebase(string: String): Int {
+    override suspend fun getPriceFirebase(string: String): Float {
         return firebaseDatabase.getPriceFirebase(string = string)
     }
