+++ app/src/main/java/team/four/mys/domain/usecases/GetPriceFireBaseUseCase.kt
@@ -4,7 +4,7 @@ import team.four.mys.domain.repository.FirebaseRepository
 
 class GetPriceFireBaseUseCase(private val firebaseRepository: FirebaseRepository) {
 
-    suspend fun execute(string: String): Int {
+    suspend fun execute(string: String): Float {
         return firebaseRepository.getPriceFirebase(string = string)
     }
 }
